Add surplus item to each order in recent savings (#2301)
* Add surplus item to each order in recent savings

* PR comments and fix

* Refactor
