+++ src/custom/components/OrderProgressBar/index.tsx
@@ -32,7 +32,7 @@ import { getExplorerOrderLink } from 'utils/explorer'
 import { useIsSmartContractWallet } from '@cow/common/hooks/useIsSmartContractWallet'
 import { useCancelOrder } from '@cow/common/hooks/useCancelOrder'
 import { CancelButton } from '@cow/common/pure/CancelButton'
-import { getExecutedSummaryData } from '@src/custom/state/orders/helpers'
+import { getExecutedSummaryData } from '@cow/utils/getExecutedSummaryData'
 import { Order } from 'state/orders/actions'
 import styled from 'styled-components/macro'
 import { DisplayLink } from '../TransactionConfirmationModal'
