+++ src/custom/state/orders/middleware.ts
@@ -17,7 +17,7 @@ import { isOrderInPendingTooLong, openNpsAppziSometimes } from 'utils/appzi'
 import { OrderObject, OrdersStateNetwork } from 'state/orders/reducer'
 import { timeSinceInSeconds } from '@cow/utils/time'
 import { getExplorerOrderLink } from 'utils/explorer'
-import { getExecutedSummary } from './helpers'
+import { getExecutedSummary } from '@cow/common/pure/ExecutedSummary'
 import { Order } from 'state/orders/actions'
 
 // action syntactic sugar
